requirements.txt
pillow
opencv-python-headless
numpy
collections